Clean Up Right-Click Menus
The context menu brought up when right-clicking in the cork is now cleaner and more condensed. Actions are now created when the cork is, instead of after a click. Actions are hidden and shown based on the type of note click.
TODO: fix board link note creation through the right click menu, delete all actions and menus when the cork is deleted.
:dragon: